Extract options to separate directory
